set up autoconversion from S3
